question progress on 8 out of 9

+++ projects/cleveland-clinic_sql-test/sql_table_definitions/Encounter.sql
@@ -1,9 +1,10 @@
-CREATE TABLE Encounter (PAT_ENC_CSN_ID DOUBLE, 
+CREATE TABLE Encounter (
+    PAT_ENC_CSN_ID INTEGER,
     ED_Arrival_Time DATE,
-    Means_Of_Arrival_CD DOUBLE,
+    Means_Of_Arrival_CD INTEGER,
     ED_Department VARCHAR,
     ED_Departure_Time DATE,
     Attending_Name VARCHAR,
     Treatment_Team_Name VARCHAR,
-    Age DOUBLE
+    Age INTEGER
 );
